Move plan execution to PlanNode
